Don't use jsdom private API features to get global properties
Before we were manually grabbing `window._core` which was a bit of a hack. This sets the correct settings to allow us to get all globals with `Object.getOwnPropertyNames(window)`.

We also need to clone the settings object before passing it to jsdom otherwise it tacks loads of stuff on to the referenced object.